Renamed _bias.py -> _compare_with_refs.py
